Updated Layout changes

+++ Marbale.Business/TransactionBL.cs
@@ -243,7 +243,7 @@ namespace Marble.Business
             trxData.ConsolidateCards(toCard, lastUpdatedBy);
         }
 
-        public void RefundCard(int cardId, decimal refundAmount, decimal credits, int faceValue, bool valid, string lastUpdatedBy)
+        public void RefundCard(int cardId, decimal refundAmount, decimal credits, decimal faceValue, bool valid, string lastUpdatedBy)
         {
             trxData.RefundCard(cardId, refundAmount, credits, faceValue, valid, lastUpdatedBy);
         }
